+++ client/actions/ActionTypes.ts
@@ -1,0 +1,15 @@
+// Types
+export const AuthTypes = {
+  AUTH_USER: 'AUTH_USER',
+  LOGIN_USER: 'LOGIN_USER',
+  REGISTER_USER: 'REGISTER_USER',
+  UNAUTH_USER: 'UNAUTH_USER',
+  AUTH_ERROR: 'AUTH_ERROR'
+}
+
+export const TodoTypes = {
+  TODO_CREATE: 'TODO_CREATE',
+  TODO_LOAD: 'TODO_LOAD',
+  TODO_LOAD_ALL: 'TODO_LOAD_ALL',
+  TODO_DELETE: 'TODO_DELETE'
+}
