+++ client/components/index.tsx
@@ -1,0 +1,5 @@
+export {default as NotFound} from '../scenes/NotFound/NotFound';
+export {default as Field} from './Field/Field';
+export {default as PasswordField} from './PasswordField/PasswordField';
+export {default as Toast} from './Toast/Toast';
+export {default as Navigation} from './Navigation/Navigation';
