+++ client/templates/Register/Register.tsx
@@ -2,28 +2,30 @@ import React from 'react'
 import { connect } from 'react-redux'
 import { Field, reduxForm, InjectedFormProps } from 'redux-form'
 
-import { loginUser, LoginInfo } from '@/modules/auth'
-import Toast from '@/components/Toast'
-import FieldComponent from '@/components/Fields/Field'
-import PasswordFieldComponent from '@/components/Fields/PasswordField'
+import { registerUser, RegisterInfo } from '@/actions/User'
+import { 
+  Toast,
+  Field as FieldComponent, 
+  PasswordField 
+} from '@/components'
 
 const required = (value: any) => (value ? undefined : 'Required')
 
 const form = reduxForm({
-  form: 'login'
+  form: 'register'
 })
 
-interface LoginProps {
-  errorMessage: string
-  loginUser: (_: LoginInfo) => any
+interface RegisterProps {
+  registerUser: (_: RegisterInfo) => any
   handleSubmit: (_: any) => any
+  errorMessage: string
 }
 
-const Login: (React.FunctionComponent<LoginProps & InjectedFormProps>) = ({ errorMessage, handleSubmit, loginUser }) => {
+const Register: React.FunctionComponent<RegisterProps & InjectedFormProps> = ({ handleSubmit, registerUser, errorMessage }) => {
   return (
     <div className='container'>
-      <h3>Login to @</h3>
-      <form onSubmit={handleSubmit((loginInfo: LoginInfo) => loginUser(loginInfo))}>
+      <h3>Register for @</h3>
+      <form onSubmit={handleSubmit((registerInfo: RegisterInfo) => registerUser(registerInfo))}>
         {errorMessage && <Toast text={errorMessage} type='error' />}
         <label htmlFor='username'>Username</label>
         <Field
@@ -37,25 +39,23 @@ const Login: (React.FunctionComponent<LoginProps & InjectedFormProps>) = ({ erro
         <Field
           id='password'
           name='password'
-          component={PasswordFieldComponent}
+          component={PasswordField}
           type='password'
           validate={required}
         />
         <button type='submit' className='btn'>
-          Login
+          Register
         </button>
       </form>
     </div>
   )
 }
 
-function mapStateToProps (state: any) {
-  return {
-    errorMessage: state.auth.error,
-    message: state.auth.message
-  }
-}
+const mapStateToProps = (state: any) => ({
+  errorMessage: state.auth.error,
+  message: state.auth.message
+})
 
-export default connect(
-  mapStateToProps, { loginUser }
-)(form(Login))
+const mapDispatchToProps = { registerUser }
+
+export default connect(mapStateToProps, mapDispatchToProps)(form(Register))
