+++ client/templates/index.tsx
@@ -1,0 +1,6 @@
+export {default as Home} from './Home/Home';
+export {default as Todos} from './Todos/Todos';
+export {default as Todo} from './Todo/Todo';
+export {default as Login} from './Login/Login';
+export {default as Register} from './Register/Register';
+export {default as NotFound} from './NotFound/NotFound';
